+++ models/encoding/src/Int32Data.cpp
@@ -65,7 +65,7 @@ Int32Data::Int32Data(
       }
    }
 
-   int const ptr_i32_size = 5 + ( ( offset < 10 ) ? offset : 10 );
+   int const ptr_i32_size = (int)( 5 + ( ( offset < 10 ) ? offset : 10 ) );
 
    ptr_i32 = static_cast< int * >( TMM_declare_var_1d( "int", ptr_i32_size ) );
    for ( int i = 0; i < ptr_i32_size; ++i ) {
